Create Task-1
Input data: The number is received at the program input 𝑛 (1<=𝑛<=100) – the size of the matrix, and then 𝑛 rows by 𝑛 numbers, each of which is equal to 0 or 1 – the matrix itself. Output data: Print "YES" if the reduced matrix can be the adjacency matrix of a
simple undirected graph, and "NO" otherwise.